🩹 fix 일정 작성시 인증확인,일정삭제시 자격확인
 일정 작성시 로그인해야만
 일정을 작성 할 수 있도록 수정
 일정 삭제시 본인의 일정만 삭제 할 수 있도록 수정

+++ src/main/java/com/ch3schedulerdevelop/scheduler/service/SchedulerService.java
@@ -21,8 +21,8 @@ public class SchedulerService {
     private final UserRepository userRepository;
 
     @Transactional
-    public CreateSchedulerResponse saveScheduler(CreateSchedulerRequest request){
-        User user = userRepository.findById(request.getUserId()).orElseThrow(
+    public CreateSchedulerResponse saveScheduler(Long sessionUserId, CreateSchedulerRequest request){
+        User user = userRepository.findById(sessionUserId).orElseThrow(
                 () -> new IllegalArgumentException("없는 유저입니다.")
         );
         Scheduler scheduler = new Scheduler(
@@ -70,10 +70,13 @@ public class SchedulerService {
     }
 
     @Transactional
-    public void deleteScheduler(Long schedulerId){
+    public void deleteScheduler(Long schedulerId, Long sessionUserId){
         Scheduler scheduler = schedulerRepository.findById(schedulerId).orElseThrow(
                 () -> new IllegalStateException("없는 일정입니다.")
         );
+        if (!scheduler.getUser().getId().equals(sessionUserId)){
+            throw new IllegalStateException("본인의 일정만 수정할 수 있습니다.");
+        }
         schedulerRepository.delete(scheduler);
     }
 }
